requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.1
transformers==4.35.2
